Refactor : todolist, othercadetlist 컴포넌트 분리 / index.js 정상화 / isFutureTodo.js 파일 이름 변경
